Updating PlayerWrapper and fixing vanilla commands
This commit resolves CoolLord22/OtherDrops#27 by trying to run the
dispatch command. The issue arose when the player couldn't be cast as a
CraftPlayer, so now we will try to run the command and if it catches an
exception, we will handle by creating a craftplayer which will execute
the command.